+++ changelog.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## Upcoming
+* Added block data
+  * In the `data/block_data` folder can now be placed JSON to modify blocks properties, such as hardness, explosion resistance, speed and jump factor and a few more
+  * With this, harder crops json has been removed and moved to a brand new data pack (block_data) that also changes more stuff
+    * To make crops able to be broken faster by hoes (and not axes) they must be added in the `iguanatweaksreborn:harder_crops` block tag
 * Falling in a boat will now always break the boat and passenger will take fall damage
 
 ## 3.6.2